Some type changes (read long description)
Changed ids and indexes from usize to u32, changed some type aliases to
structs, implemented serializing and deserializing to a json map key (a
number which gets converted to a string) for Socket IDs. Implemented
serde traits for ModulePath.

+++ src/stdlib/bool_node.rs
@@ -39,7 +39,7 @@ impl Object for bool {
 }
 
 impl Node for BoolConstructor {
-    fn execute(&self, context: &mut ExecutionContext) -> usize {
+    fn execute(&self, context: &mut ExecutionContext) -> u32 {
         let cond = context.get_inputs()[0].as_bool();
         context.set_outputs(vec![Rc::new(cond) as Rc<dyn Object>]);
         0
